@@ -19,8 +19,8 @@
   "license": "ISC",
   "dependencies": {
     "mobile-drag-drop": "^2.3.0-rc.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.12",
